requirements.txt
litellm
python-dotenv
google-adk